- fix synchronized scope - fix to execute query in Stream processing
