+++ src/main/java/jp/co/future/uroborosql/SqlAgentImpl.java
@@ -4,11 +4,8 @@ import java.sql.CallableStatement;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.Comparator;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
+import java.util.*;
+import java.util.function.Consumer;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
@@ -20,6 +17,7 @@ import jp.co.future.uroborosql.converter.MapResultSetConverter;
 import jp.co.future.uroborosql.converter.ResultSetConverter;
 import jp.co.future.uroborosql.exception.EntitySqlRuntimeException;
 import jp.co.future.uroborosql.exception.OptimisticLockException;
+import jp.co.future.uroborosql.exception.UroborosqlSQLException;
 import jp.co.future.uroborosql.filter.SqlFilterManager;
 import jp.co.future.uroborosql.mapping.EntityHandler;
 import jp.co.future.uroborosql.mapping.MappingUtils;
@@ -192,7 +190,45 @@ public class SqlAgentImpl extends AbstractAgent {
 	 */
 	@Override
 	public <T> Stream<T> query(final SqlContext sqlContext, final ResultSetConverter<T> converter) throws SQLException {
-		return StreamSupport.stream(new ResultSetSpliterator<T>(query(sqlContext), converter), false);
+		Stream<T> stream = StreamSupport.stream(new Spliterators.AbstractSpliterator<T>(Long.MAX_VALUE,
+				Spliterator.ORDERED | Spliterator.CONCURRENT) {
+			private ResultSet rs;
+
+			@Override
+			public boolean tryAdvance(final Consumer<? super T> action) {
+				try {
+					if (rs == null) {
+						rs = query(sqlContext);
+					}
+					if (!rs.next()) {
+						rs.close();
+						return false;
+					}
+					action.accept(converter.createRecord(rs));
+					return true;
+				} catch (RuntimeException | Error ex) {
+					try {
+						if (rs != null && !rs.isClosed()) {
+							rs.close();
+						}
+					} catch (SQLException e) {
+						e.printStackTrace();
+					}
+					throw ex;
+				} catch (SQLException ex) {
+					try {
+						if (rs != null && !rs.isClosed()) {
+							rs.close();
+						}
+					} catch (SQLException e) {
+						e.printStackTrace();
+					}
+					throw new UroborosqlSQLException(ex);
+				}
+			}
+		}, false);
+
+		return stream;
 	}
 
 	/**
