+++ src/main/java/jp/co/future/uroborosql/connection/DataSourceConnectionSupplierImpl.java
@@ -76,10 +76,13 @@ public class DataSourceConnectionSupplierImpl implements ConnectionSupplier {
 	 * @see jp.co.future.uroborosql.connection.ConnectionSupplier#getConnection(java.lang.String)
 	 */
 	@Override
-	public synchronized Connection getConnection(final String datasourceName) {
+	public Connection getConnection(final String datasourceName) {
 		try {
 			DataSource ds = dsMap.computeIfAbsent(datasourceName, DataSourceConnectionSupplierImpl::getNewDataSource);
-			Connection connection = ds.getConnection();
+			final Connection connection;
+			synchronized (ds) {
+				connection = ds.getConnection();
+			}
 			boolean autoCommit = getAutoCommit(datasourceName);
 			if (connection.getAutoCommit() != autoCommit) {
 				connection.setAutoCommit(autoCommit);
